# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API de proyecto final Desarrollo de Aplicaciones Web."
  version: "1.0.0"
  title: "Next Up Games API"
host: "localhost:8080"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /interes/create:
    post:
      tags:
      - "interes"
      summary: "Crear interes"
      description: ""
      operationId: "createInterest"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objecto interes creado"
        required: true
        schema:
          $ref: "#/definitions/Interes"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Interes"
  /interes/delete:
    delete:
      tags:
      - "interes"
      summary: "Borrar interes"
      description: ""
      operationId: "deleteInterest"
      produces:
      - "application/json"
      parameters:
      - name: "idUsuario"
        in: "header"
        description: "El id del usuario del interes que tiene que ser eliminado"
        required: true
        type: "integer"
      - name: "idJuego"
        in: "header"
        description: "El id del juego del interes que tiene que ser eliminado"
        required: true
        type: "integer"
      responses:
        400:
          description: "Invalid interest id supplied"
        404:
          description: "Interest not found"
      x-swagger-router-controller: "Interes"
  /interes/getAll:
    get:
      tags:
      - "interes"
      summary: "Recibir todos los intereses"
      description: ""
      operationId: "getAllInterests"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Interes"
        404:
          description: "Interest not found"
      x-swagger-router-controller: "Interes"
...
  /juegos_genero/create:
    post:
      tags:
      - "juegos_genero"
      summary: "Crear juegos_genero"
      description: ""
      operationId: "createGame_Genre"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objecto juegos_genero creado"
        required: true
        schema:
          $ref: "#/definitions/Juegos_genero"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Juegosgenero"
  /juegos_genero/delete:
    delete:
      tags:
      - "juegos_genero"
      summary: "Borrar juegos_genero"
      description: ""
      operationId: "deleteGame_Genre"
      produces:
      - "application/json"
      parameters:
      - name: "idJuego"
        in: "header"
        description: "id del juego del juegos_genero que tiene que ser eliminado"
        required: true
        type: "integer"
      - name: "idGenero"
        in: "header"
        description: "id del genero del juegos_genero que tiene que ser eliminado"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid game_genre id supplied"
        404:
          description: "Game_genre not found"
      x-swagger-router-controller: "Juegosgenero"
  /juegos_genero/getAll:
    get:
      tags:
      - "juegos_genero"
      summary: "Recibir todos los valores de juegos_generos"
      description: ""
      operationId: "getAllGame_Genres"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Juegos_genero"
        404:
          description: "Game_Genre not found"
      x-swagger-router-controller: "Juegosgenero"
  /juegos_genero/getAllGameByGenreId:
    get:
      tags:
      - "juegos_genero"
      summary: "Recibir todos los juegos de un genero por id"
      description: ""
      operationId: "getAllGameByGenreId"
      produces:
      - "application/json"
      parameters:
      - name: "idGenero"
        in: "header"
        description: "El id del genero del que se quiere recibir juegos."
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Juegos_genero"
        400:
          description: "Invalid Game_Genre id supplied"
        404:
          description: "Game_Genre not found"
      x-swagger-router-controller: "Juegosgenero"
  /juegos_genero/getAllGenreByGameId:
    get:
      tags:
      - "juegos_genero"
      summary: "Recibir todos los generos de un juego por id"
      description: ""
      operationId: "getAllGenreByGameId"
      produces:
      - "application/json"
      parameters:
      - name: "idJuego"
        in: "header"
        description: "El id del juego del que se quiere recibir sus generos."
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Juegos_genero"
        400:
          description: "Invalid Game_Genre id supplied"
        404:
          description: "Game_Genre not found"
      x-swagger-router-controller: "Juegosgenero"
  /juegos_genero/update:
    put:
      tags:
      - "juegos_genero"
      summary: "Actualizar juegos_genero"
      description: ""
      operationId: "updateGame_Genre"
      produces:
      - "application/json"
      parameters:
      - name: "idJuego"
        in: "header"
        description: "id del juego del juegos_genero que tiene que ser actualizado"
        required: true
        type: "integer"
      - name: "idGenero"
        in: "header"
        description: "id del genero del juegos_genero que tiene que ser actualizado"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Objeto interes actualizado"
        required: true
        schema:
          $ref: "#/definitions/Juegos_genero"
      responses:
        400:
          description: "Invalid game_genre id supplied"
        404:
          description: "Game_platform not found"
      x-swagger-router-controller: "Juegosgenero"
  /juegos_plataforma/create:
    post:
      tags:
      - "juegos_plataforma"
      summary: "Crear juegos_plataforma"
      description: ""
      operationId: "createGame_Platform"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objecto juegos_plataforma creado"
        required: true
        schema:
          $ref: "#/definitions/Juegos_plataforma"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Juegosplataforma"
  /juegos_plataforma/delete:
    delete:
      tags:
      - "juegos_plataforma"
      summary: "Borrar juegos_plataforma"
      description: ""
      operationId: "deleteGame_Platform"
      produces:
      - "application/json"
      parameters:
      - name: "idJuego"
        in: "header"
        description: "id del juego del juegos_plataforma que tiene que ser eliminado"
        required: true
        type: "integer"
      - name: "idPlataforma"
        in: "header"
        description: "id del genero del juegos_plataforma que tiene que ser eliminado"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid game_platform id supplied"
        404:
          description: "Game_platform not found"
      x-swagger-router-controller: "Juegosplataforma"
  /juegos_plataforma/getAll:
    get:
      tags:
      - "juegos_plataforma"
      summary: "Recibir todos los valores de juegos_plataforma"
      description: ""
      operationId: "getAllGame_Platforms"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Juegos_plataforma"
        404:
          description: "Game_PLatform not found"
      x-swagger-router-controller: "Juegosplataforma"
  /juegos_plataforma/getAllGameByPlatformId:
    get:
      tags:
      - "juegos_plataforma"
      summary: "Recibir todos los juegos de una plataforma por id"
      description: ""
      operationId: "getAllGameByPlatformId"
      produces:
      - "application/json"
      parameters:
      - name: "idPlataforma"
        in: "header"
        description: "El id de la plataforma del que se quiere recibir juegos."
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Juegos_plataforma"
        400:
          description: "Invalid game_platform id supplied"
        404:
          description: "Game_platform not found"
      x-swagger-router-controller: "Juegosplataforma"
  /juegos_plataforma/getAllPlatformByGameId:
    get:
      tags:
      - "juegos_plataforma"
      summary: "Recibir todas las plataformas de un juego por id"
      description: ""
      operationId: "getAllPlatformByGameId"
      produces:
      - "application/json"
      parameters:
      - name: "idJuego"
        in: "header"
        description: "El id del juego del que se quiere recibir sus plataformas."
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Juegos_plataforma"
        400:
          description: "Invalid Game_platform id supplied"
        404:
          description: "Game_platform not found"
      x-swagger-router-controller: "Juegosplataforma"
  /juegos_plataforma/update:
    put:
      tags:
      - "juegos_plataforma"
      summary: "Actualizar juegos_plataforma"
      description: ""
      operationId: "updateGame_Platform"
      produces:
      - "application/json"
      parameters:
      - name: "idJuego"
        in: "header"
        description: "id del juego del juegos_plataforma que tiene que ser actualizado"
        required: true
        type: "integer"
      - name: "idPlataforma"
        in: "header"
        description: "id de la plataforma del juegos_plataforma que tiene que ser\
          \ actualizado"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Objeto juegos_plataforma actualizado"
        required: true
        schema:
          $ref: "#/definitions/Juegos_plataforma"
      responses:
        400:
          description: "Invalid game_platform id supplied"
        404:
          description: "Game_platform not found"
      x-swagger-router-controller: "Juegosplataforma"
...
  /usuario/create:
    post:
      tags:
      - "usuario"
      summary: "Crear usuario"
      description: ""
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objecto usuario creado"
        required: true
        schema:
          $ref: "#/definitions/Usuario"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Usuario"
  /usuario/delete:
    delete:
      tags:
      - "usuario"
      summary: "Borrar usuario"
      description: ""
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "idUsuario"
        in: "header"
        description: "El id del usuario que tiene que ser eliminado"
        required: true
        type: "integer"
      responses:
        400:
          description: "Invalid user id supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Usuario"
  /usuario/getAll:
    get:
      tags:
      - "usuario"
      summary: "Recibir todos los usuarios"
      description: ""
      operationId: "getAllUsers"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Usuario"
        404:
          description: "User not found"
      x-swagger-router-controller: "Usuario"
  /usuario/getOneById:
    get:
      tags:
      - "usuario"
      summary: "Recibir un usuario por id"
      description: ""
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "idUsuario"
        in: "header"
        description: "El id del usuario que tiene que ser recogido."
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Usuario"
        400:
          description: "Invalid user id supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Usuario"
  /usuario/update:
    put:
      tags:
      - "usuario"
      summary: "Actualizar usuario"
      description: ""
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
      - name: "idUsuario"
        in: "header"
        description: "id del usuario que tiene que ser actualizado"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Objeto usuario actualizado"
        required: true
        schema:
          $ref: "#/definitions/Usuario"
      responses:
        400:
          description: "Invalid user id supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Usuario"
definitions:
  Usuario:
    type: "object"
    properties:
      idUsuario:
        type: "integer"
        format: "int64"
      apodo:
        type: "string"
      email:
        type: "string"
        format: "email"
      password:
        type: "string"
      imagen:
        type: "string"
    example:
      apodo: "string"
      email: "user@example.com"
      password: "string"
      imagen: "string"
...
  Interes:
    type: "object"
    properties:
      idJuego:
        $ref: "#/definitions/Juego/properties/idJuego"
      idUsuario:
        $ref: "#/definitions/Usuario/properties/idUsuario"
  Juegos_genero:
    type: "object"
    properties:
      idJuego:
        $ref: "#/definitions/Juego/properties/idJuego"
      idGenero:
        $ref: "#/definitions/Genero/properties/idGenero"
  Juegos_plataforma:
    type: "object"
    properties:
      idJuego:
        $ref: "#/definitions/Juego/properties/idJuego"
      idPlataforma:
        $ref: "#/definitions/Plataforma/properties/idPlataforma"